exo-k8s-dbaas-filter/requirements.txt
requests>=2.31.0
requests-exoscale-auth>=0.3.0
